Implement option to hide the reader view footer slider (#15)
Sometimes when im reading a book, and trying to use android gestures
(swipe up to go home) the navbar to skip pages/chapters shows up and i
fat finger that losing the page i was reading on.

This patch adds a toggle to disable the whole footer bar in reading view.

Co-authored-by: Tuxx <tuxx@danktank.nl>
Co-authored-by: Patrick Gaskin <patrick@pgaskin.net>